requirements.txt
fastapi
uvicorn
faiss-cpu
sentence-transformers
tqdm
markdownify
playwright
beautifulsoup4
